Fixed exceptions in mapper modules (#7581)
Added missing imports, fixed other issues in  marc_mapper,
hathi_mapper, and uiuc_marc_mapper.

Note that the keyErrorAsNone parameter is now passed as True to
getprop() in MARCMapper._get_mapped_value(), which should result
in an empty list if the property doesn't exists, which seemed fine
because that's a condition that's already handled in the next
statement in that function.