Add SCDLCharlestonMapper for object mapping (#7651)
SCDLCharlestonMapper extends SCDLMapper in order to perform special
mapping of the 'object' property that is different from that of
the other SCDL providers.  The included test describes the desired
behavior.

Also, change Charleston "sets" property, since Charleston have
foregone sets with their new server, and this is necessary to
be able to fetch something that can be mapped.