Replace amara.httplib2 with urllib2 in oai module
The amara.httplib2 module doesn't work with SSL on all systems,
and it's not as well-maintained as the standard library's urllib2.
Indiana is the first provider to have an SSL endpoint, so that's
why this change is included in this branch.  The fetch works on
OS X with a combination of homebrew and system libraries, but does
not work on at least one Debian wheezy/sid system; wheras urllib2
works fine in the latter case.