Exclude all network fetcher tests from Travis CI
The fetcher tests are too problematic to run as CI tests.  They do
not provide the utility that they are supposed to, produce false
errors about the correctness of the code and code breakages, and
get in the way of our workflow.

As we have discussed, testing of provider endpoints should be done
by means other than automated tests.  For the time being, the
fetcher tests will still execute in the course of normal "all tests"
runs that we perform ourselves in order to run tests requiring
authentication and network access.