Fix exception in AbsoluteURLFetcher
In `fetch_all_data`, make method calls that depend on `records` being
iterable conditionally upon it evaluating to true.  This is in-line with
what has already been done in `Fetcher.add_provider_to_item_records()`.

During a fetch of University of Virginia, we encountered a set that had
no records, apparently due to an incorrect URL being passed to us.
Rather than having the fetch fail due to an unhandled exception, we want
to iterate conditionally over `records` only if it's populated, and move
on with the fetch if it's not.