Fix collapsing of spatial objects
Fix issue caused recently by the way we passed on place names
equaling "United States" or one of its variants. This fix allows
multiple spatial entries to be rolled up into one, for example, if one
is for "name" equaling "Los Angeles" and another is for "name" equaling
"United States" we can say this is for a single place, Los Angeles,
which is in the United States.

We will take all of these (which include "États-Unis" and "USA") and
normalize them to "United States" because the new behavior after
removing Bing lookups is for places where the name equals the country
after a GeoNames lookup to not have coordinates added to the Place data.