Remove 'class geocoding' from akara config template
Remove obsolete 'class geocoding' section from akara.conf.template. I
think that this was a leftover from many years ago when a Zephiera
module was being used for the geocoder, before the `geocode' module was
added. If you search `ingestion' for 'geocoding' you will find that
this class is not used anywhere anymore.