Strip commas & spaces from creator values
We were getting creator names from one provider that were formatted with
commas, like "Strecker, Herman,". This value is then concatenated with a
date using a comma delimiter, resulting in ",," appearing in the result.
I assume it's safe for all OAIMODSMappers to strip this field in this
fashion.

See https://digitalpubliclibraryofamerica.atlassian.net/browse/DT-1191